requirements.txt
ultralytics
tensorflow>=2.14.0
numpy
Pillow
flask
